+++ common/src/main/java/io/netty/util/Recycler.java
@@ -16,13 +16,15 @@
 package io.netty.util;
 
 import io.netty.util.concurrent.FastThreadLocal;
-import io.netty.util.internal.BlockingMessageQueue;
 import io.netty.util.internal.ObjectPool;
+import io.netty.util.internal.PlatformDependent;
 import io.netty.util.internal.SystemPropertyUtil;
 import io.netty.util.internal.logging.InternalLogger;
 import io.netty.util.internal.logging.InternalLoggerFactory;
 import org.jctools.queues.MessagePassingQueue;
 
+import java.util.ArrayDeque;
+import java.util.Queue;
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 
 import static io.netty.util.internal.PlatformDependent.newMpscQueue;
@@ -285,6 +287,114 @@ public abstract class Recycler<T> {
                 return new DefaultHandle<T>(this);
             }
             return null;
+        }
+    }
+
+    /**
+     * This is an implementation of {@link MessagePassingQueue}, similar to what might be returned from
+     * {@link PlatformDependent#newMpscQueue(int)}, but intended to be used for debugging purpose.
+     * The implementation relies on synchronised monitor locks for thread-safety.
+     * The {@code drain} and {@code fill} bulk operations are not supported by this implementation.
+     */
+    private static final class BlockingMessageQueue<T> implements MessagePassingQueue<T> {
+        private final Queue<T> deque;
+        private final int maxCapacity;
+
+        BlockingMessageQueue(int maxCapacity) {
+            this.maxCapacity = maxCapacity;
+            // This message passing queue is backed by an ArrayDeque instance,
+            // made thread-safe by synchronising on `this` BlockingMessageQueue instance.
+            // Why ArrayDeque?
+            // We use ArrayDeque instead of LinkedList or LinkedBlockingQueue because it's more space efficient.
+            // We use ArrayDeque instead of ArrayList because we need the queue APIs.
+            // We use ArrayDeque instead of ConcurrentLinkedQueue because CLQ is unbounded and has O(n) size().
+            // We use ArrayDeque instead of ArrayBlockingQueue because ABQ allocates its max capacity up-front,
+            // and these queues will usually have large capacities, in potentially great numbers (one per thread),
+            // but often only have comparatively few items in them.
+            deque = new ArrayDeque<T>();
+        }
+
+        @Override
+        public synchronized boolean offer(T e) {
+            if (deque.size() == maxCapacity) {
+                return false;
+            }
+            return deque.offer(e);
+        }
+
+        @Override
+        public synchronized T poll() {
+            return deque.poll();
+        }
+
+        @Override
+        public synchronized T peek() {
+            return deque.peek();
+        }
+
+        @Override
+        public synchronized int size() {
+            return deque.size();
+        }
+
+        @Override
+        public synchronized void clear() {
+            deque.clear();
+        }
+
+        @Override
+        public synchronized boolean isEmpty() {
+            return deque.isEmpty();
+        }
+
+        @Override
+        public int capacity() {
+            return maxCapacity;
+        }
+
+        @Override
+        public boolean relaxedOffer(T e) {
+            return offer(e);
+        }
+
+        @Override
+        public T relaxedPoll() {
+            return poll();
+        }
+
+        @Override
+        public T relaxedPeek() {
+            return peek();
+        }
+
+        @Override
+        public int drain(Consumer<T> c, int limit) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public int fill(Supplier<T> s, int limit) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public int drain(Consumer<T> c) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public int fill(Supplier<T> s) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public void drain(Consumer<T> c, WaitStrategy wait, ExitCondition exit) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public void fill(Supplier<T> s, WaitStrategy wait, ExitCondition exit) {
+            throw new UnsupportedOperationException();
         }
     }
 }
